+++ backend/substrapp/views/utils.py
@@ -7,6 +7,8 @@ from wsgiref.util import is_hop_by_hop
 
 import grpc
 from django.conf import settings
+from django.db.models import Count
+from django.db.models import Q
 from django.http import FileResponse
 from rest_framework import status
 from rest_framework.authentication import BasicAuthentication
@@ -20,6 +22,7 @@ import orchestrator.computetask_pb2 as computetask_pb2
 import orchestrator.error
 import orchestrator.event_pb2 as event_pb2
 import orchestrator.model_pb2 as model_pb2
+from localrep.models import ComputeTask as ComputeTaskRep
 from node.authentication import NodeUser
 from orchestrator import client as orc_client
 from orchestrator import failure_report_pb2
@@ -378,20 +381,39 @@ def to_string_uuid(str_or_hex_uuid: uuid.UUID) -> str:
     return str(uuid.UUID(str_or_hex_uuid))
 
 
-def add_cp_status_and_task_counts(client, data):
-    compute_plan = client.query_compute_plan(data["key"])
-    return update_cp_status_and_task_counts(data, compute_plan)
+def get_cp_status(stats):
+    """
+    Compute cp status from tasks counts.
+    See: `orchestrator/lib/persistence/computeplan_dbal.go`
+    """
+    if stats["task_count"] == 0:
+        return computeplan_pb2.PLAN_STATUS_UNKNOWN
+    elif stats["done_count"] == stats["task_count"]:
+        return computeplan_pb2.PLAN_STATUS_DONE
+    elif stats["failed_count"] > 0:
+        return computeplan_pb2.PLAN_STATUS_FAILED
+    elif stats["canceled_count"] > 0:
+        return computeplan_pb2.PLAN_STATUS_CANCELED
+    elif stats["waiting_count"] == stats["task_count"]:
+        return computeplan_pb2.PLAN_STATUS_WAITING
+    elif stats["waiting_count"] < stats["task_count"] and stats["doing_count"] == 0 and stats["done_count"] == 0:
+        return computeplan_pb2.PLAN_STATUS_TODO
+    else:
+        return computeplan_pb2.PLAN_STATUS_DOING
 
 
-def update_cp_status_and_task_counts(data, compute_plan):
-    data["status"] = compute_plan["status"]
-    data["task_count"] = compute_plan["task_count"]
-    data["done_count"] = compute_plan["done_count"]
-    data["waiting_count"] = compute_plan["waiting_count"]
-    data["todo_count"] = compute_plan["todo_count"]
-    data["doing_count"] = compute_plan["doing_count"]
-    data["canceled_count"] = compute_plan["canceled_count"]
-    data["failed_count"] = compute_plan["failed_count"]
+def add_cp_status_and_task_counts(data):
+    stats = ComputeTaskRep.objects.filter(compute_plan__key=data["key"]).aggregate(
+        task_count=Count("key"),
+        done_count=Count("key", filter=Q(status=computetask_pb2.STATUS_DONE)),
+        waiting_count=Count("key", filter=Q(status=computetask_pb2.STATUS_WAITING)),
+        todo_count=Count("key", filter=Q(status=computetask_pb2.STATUS_TODO)),
+        doing_count=Count("key", filter=Q(status=computetask_pb2.STATUS_DOING)),
+        canceled_count=Count("key", filter=Q(status=computetask_pb2.STATUS_CANCELED)),
+        failed_count=Count("key", filter=Q(status=computetask_pb2.STATUS_FAILED)),
+    )
+    data.update(stats)
+    data["status"] = computeplan_pb2.ComputePlanStatus.Name(get_cp_status(stats))
     return data