+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Local representation of computetask assets (#688)
 - Add `ORCHESTRATOR_RABBITMQ_ACTIVITY_TIMEOUT` to restart event app if no activity (#739)
 - Exponential backoff of celery tasks retry (#736)
+- Compute CP tasks count and status from localrep data (#717)
 
 ### Changed
 - When resyncing the local representation, only fetch the latest events (#656)